Restructure project layout
- Create directories, such as build, scripts, web
- Relocate operation objects and those functions to internal
- Relocate data objects and those functions to internal/data-object
- Relocate test app (pitcher and catcher) to internal/app
