+++ poc-cb-net/internal/data-objects/network-interface.go
@@ -1,14 +1,12 @@
 package dataobjects
 
-
-type IP struct{
+type IP struct {
 	Version   string `json:"Version"`
 	IPAddress string `json:"IP"`
 	CIDRBlock string `json:"CIDRBlock"`
 }
 
-type NetworkInterface struct{
-	Name	string  `json:"Name"`
-	IPs		[]IP 	`json:"IPs"`
+type NetworkInterface struct {
+	Name string `json:"Name"`
+	IPs  []IP   `json:"IPs"`
 }
-
