+++ poc-cb-net/internal/data-objects/networking-rule.go
