+++ poc-cb-net/internal/data-objects/vm-network-information.go
