+++ poc-cb-net/internal/private-ip-checker.go
@@ -1,4 +1,4 @@
-package poc_cb_net
+package internal
 
 import (
 	"fmt"
